layout/generic/test/test_bug1460740.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>
<style>
  @font-face {
    font-family: "Ahem";
    src: url(/tests/fonts/Ahem.ttf);
  }
  details {
    font: 16px/1 Ahem;
  }
</style>
<details id="d"></details>
<script>
  const selection = getSelection();
  d.addEventListener("DOMSubtreeModified", ev => {
    window.find("cthulu", true, false);
    selection.extend(document.body, document.body.childNodes.length);
  })

  SimpleTest.waitForExplicitFinish();
  window.onload = () => {
    selection.collapse(document.body);
    synthesizeMouse(d, 20, 10, { type: "mousedown" });
    synthesizeMouse(d, 20, 10, { type: "mouseup" });
    is(selection.focusNode, document.body);
    SimpleTest.finish();
  };
</script>
